% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_blogs.R
\name{initiate_blog}
\alias{initiate_blog}
\title{Addin Application to set up required pharmaverse blog folders/files}
\usage{
initiate_blog()
}
\description{
Addin Application to set up required pharmaverse blog folders/files
}
