% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmaverse4devs-package.R
\docType{package}
\name{pharmaverse4devs-package}
\alias{pharmaverse4devs}
\alias{pharmaverse4devs-package}
\title{pharmaverse4devs: Storing Utility Functions and Development Tools for Packages Across pharmaverse}
\description{
Storing Utility Functions and Development Tools for Packages Across pharmaverse.
}
\author{
\strong{Maintainer}: Sadchla Mascary \email{sadchla.mathieu@gmail.com}

}
