% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_compare_package_sizes.R
\name{run_compare_package_sizes}
\alias{run_compare_package_sizes}
\title{Open an addin to compare package file sizes}
\usage{
run_compare_package_sizes()
}
\description{
Open an addin to compare package file sizes
}
